Integrate Ophim API and refactor DAO structure
Added Retrofit, Gson, OkHttp, and Glide dependencies. Introduced OphimApiService and RetrofitClient for network calls to fetch movies. Added MovieRepository and OphimJSON models for API data mapping and local storage. Refactored all DAO classes into a dao package and updated imports accordingly. Changed movie rating type from Int to Double in database and models. Updated MainActivity to support loading and displaying movies from the API. Minor layout update to add bannerImage ImageView.

+++ app/src/main/java/com/example/android_movie_app/DatabaseHelper.kt
@@ -38,7 +38,7 @@ class DatabaseHelper(context: Context) :
                 posterUrl TEXT,
                 year INTEGER,
                 viewCount INTEGER DEFAULT 0,
-                rating INTEGER DEFAULT 0,  -- 1-5
+                rating DOUBLE DEFAULT 0,  -- 1-5
                 createdAt TEXT DEFAULT CURRENT_TIMESTAMP
             )
         """)
